+++ activerecord/test/cases/associations/belongs_to_associations_test.rb
@@ -1131,12 +1131,6 @@ class BelongsToAssociationsTest < ActiveRecord::TestCase
     Column.create! record: record
     assert_equal 1, Column.count
   end
-
-  def test_association_force_reload_with_only_true_is_deprecated
-    client = Client.find(3)
-
-    assert_deprecated { client.firm(true) }
-  end
 end
 
 class BelongsToWithForeignKeyTest < ActiveRecord::TestCase